Added args argument to make_inpaths function Fixed bug checking energy as int instead of float

+++ ez_ufo_qt/GUI/config.py
@@ -569,7 +569,7 @@ class ConfigGroup(QGroupBox):
             raise InvalidInputError("Value out of range for: SNR")
 
         # Photon energy: e_energy
-        if int(parameters.params['e_energy']) < 0:
+        if float(parameters.params['e_energy']) < 0:
             raise InvalidInputError("Value out of range for: Photon energy [keV]")
 
         # Pixel size: e_pixel
